% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{foo}
\alias{foo}
\title{Compose a Message with an Entity Name and Age}
\usage{
foo(name, age)
}
\arguments{
\item{name}{(\code{character}) The name of the entity.}

\item{age}{(\code{numeric}) The age of the entity.}
}
\value{
(\code{character}) A string with the entity's details.
}
\description{
Given \code{name} and \code{age}, when \code{foo} is
called, then a string with that entity details is composed and returned.
}
\examples{
\dontrun{
msg <- foo('Earth', 4.543e9)
message(msg)
}

}
